+++ lib/pages/live_room_page.dart
@@ -2,10 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:spotify_polls/widgets/control_bar.dart';
 import 'package:spotify_polls/widgets/song_cards.dart';
 import 'package:spotify_polls/widgets/custom_app_bar.dart';
-import 'package:spotify_polls/song_queue.dart';
+import 'package:spotify_polls/widgets/song_queue.dart';
 import 'package:spotify_polls/widgets/media_items.dart';
-
-import '../widgets/voting.dart';
+import 'package:spotify_polls/widgets/voting.dart';
 
 class LiveRoomPage extends StatefulWidget {
   const LiveRoomPage({super.key, this.title = "Live Room"});
